+++ lib/jazzy/doc_builder.rb
@@ -160,8 +160,7 @@ module Jazzy
     end
 
     def self.copy_assets(destination)
-      origin = Pathname(__FILE__).parent + '../../lib/jazzy/assets/.'
-      FileUtils.cp_r(origin, destination)
+      FileUtils.cp_r(Config.instance.assets_directory.children, destination)
       Pathname.glob(destination + 'css/**/*.scss').each do |scss|
         contents = scss.read
         css = Sass::Engine.new(contents, syntax: :scss).render
@@ -284,6 +283,7 @@ module Jazzy
       doc[:name] = doc_model.name
       doc[:kind] = doc_model.type.name
       doc[:dash_type] = doc_model.type.dash_type
+      doc[:declaration] = doc_model.declaration
       doc[:overview] = Jazzy.markdown.render(doc_model.overview)
       doc[:structure] = source_module.doc_structure
       doc[:tasks] = render_tasks(source_module, doc_model.children)